[FP-40] Feat: AlarmInfo 컴포넌트 동적으로 변경
